Restructure into src/, port SDL2->SDL3, swap SFML audio for SDL3_mixer
Restructure: all 67 source files moved into src/ (mirrors Dark-Ages).

SDL2->SDL3 port, mechanical throughout:
- CDisplay::init() rewritten for SDL3's display-enumeration API
  (SDL_GetDisplays/SDL_GetFullscreenDisplayModes/SDL_GetDesktopDisplayMode
  replace SDL_GetNumDisplayModes/SDL_GetDisplayMode/SDL_GetCurrentDisplayMode),
  window/renderer creation signatures, and SDL_Init's bool return.
- Fixed a real SDL3 behavior change this surfaced: a window can't have both
  a surface (SDL_GetWindowSurface) and a renderer at the same time, so the
  pixel format now comes from SDL_GetWindowPixelFormat instead.
- Deleted the DA2DX.cpp/.h legacy DirectDraw comment block while in there
  (fully dead, confirmed earlier - see prior commits/discussion).
- CGraphic/DA2Game's surface and pixel-format calls updated (SDL_ConvertSurface,
  SDL_SetSurfaceColorKey, SDL_CreateSurface, DA2_GetRGBA wrapper for the
  SDL_PixelFormatDetails change, per-texture SDL_SetTextureScaleMode replacing
  the removed global nearest-neighbor hint).
- DA2Input's event handling updated for the renamed event types, the
  keysym->key struct field, uppercased SDLK_* letter names, and
  SDL_GetMouseState's switch to float coordinates.
- Added src/SDLCompat.h: SDL_RenderCopy/RenderFillRect/RenderDrawRect all
  moved from int-based SDL_Rect to float-based SDL_FRect in SDL3. Rather than
  touch each of the ~150 call sites individually, these keep the old
  SDL2-shaped signatures (DA2_RenderCopy etc.) and convert at the boundary.

Audio: replaced SFML (sf::Music) with SDL3_mixer's newer MIX_ API, matching
the pattern already proven in Dark-Ages' CMusic (same vendored SDL3_mixer
build, already confirmed there to play .ogg files). This maps more directly
than classic SDL_mixer would: MIX_Track/MIX_SetTrackGain give each of the six
song slots independent gain, so the existing per-frame crossfade math in
CheckState() ports over almost line-for-line instead of needing a redesign.
cDA2Music::Init() now does real work (MIX_Init/MIX_CreateMixerDevice) and is
actually called from main(), unlike before.

Vendored third_party/sdl3 (SDL3 + SDL3_mixer static libs) copied from the
Dark-Ages repo's own third_party/sdl3; SDL2/SFML dev packages removed, no
longer needed. Fully static build - no runtime DLLs.

Verified: builds clean, runs stably, real window opens (Release only, per
the user's standing preference).

+++ src/DA2Input.h
@@ -1,7 +1,7 @@
 #ifndef _DA2INPUT_H
 #define _DA2INPUT_H
 
-#include <SDL.h>
+#include <SDL3/SDL.h>
 #include <stdio.h>
 
 enum DA2KEYS{
@@ -75,8 +75,8 @@ public:
 
 	bool keyLocked[256];
 	bool mouseLocked[2];
-	int mousePosX;
-	int mousePosY;
+	float mousePosX;
+	float mousePosY;
   int mouseX;
   int mouseY;
   float multPosX;
